requirements.txt
playwright~=1.52.0
pytest~=8.3.0
pydantic~=2.11.0
pydantic-settings~=2.9.0
allure-python-commons~=2.14.0
allure-pytest~=2.14.0
pytest-xdist~=3.0
pytest-playwright~=0.7.0
tenacity~=9.1.2